// src/features/theme/types/colors.ts
export interface RawColors {
  white?: string;
  black?: string;
  primary?: string;
  secondary?: string;
  secondary2?: string;
  secondary3?: string;
  danger?: string;
  success?: string;
  info?: string;
  warning?: string;
  border?: string;
  gray?: string;
  dark?: string;
  light?: string;
}